renderer controllers moved to api folder
